Realize logger for metrics
