+++ code/distance-between-vectors/Infrastructure/src/main/java/ru/unn/agile/infrastructure/TxtLogger.java
@@ -1,0 +1,69 @@
+package ru.unn.agile.infrastructure;
+
+import ru.unn.agile.infrastructure.viewmodel.ILogger;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.FileReader;
+import java.io.FileWriter;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.List;
+import java.util.Locale;
+
+public class TxtLogger implements ILogger {
+    private static final String DATE_FORMAT_NOW = "yyyy-MM-dd HH:mm:ss";
+    private final BufferedWriter buffWriter;
+    private final String filename;
+
+    private static String isDateNow() {
+        Calendar calendar = Calendar.getInstance();
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat(DATE_FORMAT_NOW, Locale.ENGLISH);
+        return simpleDateFormat.format(calendar.getTime());
+    }
+
+    public TxtLogger(final String filename) {
+        this.filename = filename;
+
+        BufferedWriter logWriter = null;
+        try {
+            logWriter = new BufferedWriter(new FileWriter(filename));
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        buffWriter = logWriter;
+    }
+
+    @Override
+    public void log(final String s) {
+        try {
+            buffWriter.write(isDateNow() + " > " + s);
+            buffWriter.newLine();
+            buffWriter.flush();
+
+        } catch (Exception ex) {
+            System.out.println(ex.getMessage());
+        }
+    }
+
+    @Override
+    public List<String> getLog() {
+        BufferedReader reader;
+        ArrayList<String> log = new ArrayList<String>();
+        try {
+            reader = new BufferedReader(new FileReader(filename));
+            String str = reader.readLine();
+
+            while (str != null) {
+                log.add(str);
+                str = reader.readLine();
+            }
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+        }
+
+        return log;
+    }
+
+}
