+++ code/distance-between-vectors/Infrastructure/src/test/java/ru/unn/agile/infrastructure/ViewModelWithTxtLoggerTests.java
@@ -1,0 +1,14 @@
+package ru.unn.agile.infrastructure;
+
+import ru.unn.agile.infrastructure.viewmodel.ViewModel;
+import ru.unn.agile.infrastructure.viewmodel.ViewModelTests;
+
+public class ViewModelWithTxtLoggerTests extends ViewModelTests {
+
+    @Override
+    public void setUp() {
+        TxtLogger realLogger = new TxtLogger("metricslog.txt");
+        ViewModel viewModel = new ViewModel(realLogger);
+        super.setExternalViewModel(viewModel);
+    }
+}
