--- code/distance-between-vectors/View/src/main/resources/ru/unn/agile/distancebetweenvector/view/CalculateMetrics.fxml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-<?import ru.unn.agile.distancebetweenvector.viewmodel.*?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.unn.agile.distancebetweenvector.view.CalculateMetrics">
-   <fx:define>
-      <ViewModel fx:id="viewModel" />
-   </fx:define>
-   <children>
-      <TextField  fx:id="coordX" layoutX="202.0" layoutY="43.0" prefHeight="25.0" prefWidth="31.0" />
-      <Label  fx:id="L1" layoutX="199.0" layoutY="124.0" prefHeight="25.0" prefWidth="140.0" text="${viewModel.resultL1}"/>
-      <TextField  fx:id="coordY" layoutX="269.0" layoutY="43.0" prefHeight="25.0" prefWidth="31.0" />
-      <Label layoutX="254.0" layoutY="43.0" prefHeight="25.0" prefWidth="8.0" text="y" />
-      <TextField  fx:id="coordZ" layoutX="327.0" layoutY="43.0" prefHeight="25.0" prefWidth="31.0" />
-      <Label layoutX="313.0" layoutY="43.0" prefHeight="25.0" prefWidth="8.0" text="z" />
-      <Label layoutX="77.0" layoutY="43.0" prefHeight="25.0" prefWidth="96.0" text="Введите вектор:" />
-      <Label layoutX="84.0" layoutY="124.0" prefHeight="25.0" prefWidth="77.0" text="Метрика L1" />
-      <Label layoutX="84.0" layoutY="166.0" prefHeight="25.0" prefWidth="77.0" text="Метрика L2" />
-      <Label layoutX="84.0" layoutY="210.0" prefHeight="25.0" prefWidth="77.0" text="Метрика L3" />
-      <Label layoutX="84.0" layoutY="260.0" prefHeight="25.0" prefWidth="77.0" text="Метрика L4" />
-      <Label layoutX="84.0" layoutY="307.0" prefHeight="25.0" prefWidth="77.0" text="Метрика Linf" />
-      <Label fx:id="L2" layoutX="198.0" layoutY="166.0" prefHeight="25.0" prefWidth="140.0" text="${viewModel.resultL2}"/>
-      <Label fx:id="L3" layoutX="199.0" layoutY="210.0" prefHeight="25.0" prefWidth="140.0" text="${viewModel.resultL3}"/>
-      <Label fx:id="L4" layoutX="198.0" layoutY="260.0" prefHeight="25.0" prefWidth="140.0" text="${viewModel.resultL4}"/>
-      <Label fx:id="Linf" layoutX="198.0" layoutY="307.0" prefHeight="25.0" prefWidth="140.0" text="${viewModel.resultLinf}"/>
-      <Label fx:id="message" layoutX="384.0" layoutY="37.0" prefHeight="37.0" prefWidth="171.0" text="${viewModel.message}"/>
-      <Button fx:id="btnCalc" defaultButton="true" layoutX="408.0" layoutY="299.0" mnemonicParsing="false" prefHeight="67.0" prefWidth="156.0" text="Рассчитать" textFill="#2d19ab" underline="true">
-         <font>
-            <Font size="23.0" />
-         </font>
-      </Button>
-   </children>
-</AnchorPane>
