+++ code/distance-between-vectors/ViewModel/src/main/java/ru/unn/agile/infrastructure/viewmodel/ILogger.java
@@ -1,0 +1,9 @@
+package ru.unn.agile.infrastructure.viewmodel;
+
+import java.util.List;
+
+public interface ILogger {
+    void log(String s);
+
+    List<String> getLog();
+}
